requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asttokens==2.4.1
basemap==2.0.0
basemap_data==2.0.0
basemap_data_hires==2.0.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.3.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==4.4.2
distro==1.9.0
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
gdown==5.2.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==1.2.7
langchain-core==1.2.7
langchain-groq==1.1.1
langchain-openai==1.1.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
moviepy==1.0.3
mutagen==1.47.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.1.2
openai==2.16.0
opencv-python==4.10.0.84
orjson==3.11.6
ormsgpack==1.12.2
packaging==24.1
pafy==0.5.5
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pycryptodomex==3.21.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.2.0
pyproj==3.7.1
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pywin32==308
PyYAML==6.0.3
pyzmq==26.2.0
regex==2026.1.15
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tavily-python==0.7.21
tenacity==9.1.2
textblob==0.19.0
tiktoken==0.12.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.2.3
uuid_utils==0.14.0
wcwidth==0.2.13
websockets==13.1
xxhash==3.6.0
youtube-dl==2021.12.17
yt-dlp==2024.10.7
zstandard==0.25.0
